ABMC Baja y su test corregidos

+++ src/main/java/ABMC/POI_ABMC.java
@@ -35,7 +35,7 @@ public class POI_ABMC {
 		}
 	}
 
-	public boolean delete(double ID) {
+	public boolean delete(int ID) {
 		if (DB_Server.getPOIbyId(ID) != null) {
 			return DB_Server.eliminarPOI(ID);
 		} else
